LayoutTests/svg/animations/animate-calcMode-spline-from-by-expected.txt

SVG 1.1 dynamic animation tests

Test calcMode spline with from-by animation. You should see a green 100x100 rect and only PASS messages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.x.animVal.value is 100
PASS rect.x.baseVal.value is 100
PASS rect.x.animVal.value is almost 18.8, after first half
PASS rect.x.baseVal.value is almost 18.8, after first half
PASS rect.x.animVal.value is almost 0, just before-end
PASS rect.x.baseVal.value is almost 0, just before-end
PASS rect.x.animVal.value is 100
PASS rect.x.baseVal.value is 100
PASS successfullyParsed is true

TEST COMPLETE
